Fixed two archers spawned at the start of the game
Changed how the camera switching blueprint works. the player now spawns in as an uncontrollable pawn and is prompted to press the camera switch button in order to start the game.
Also fixed an issue where the rat IMC did not get applied when swapping to the rat, and instead just inherited the archer IMC